requirements.txt
numpy==1.12.0
olefile==0.44
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
